+++ packages/jaspr/lib/browser.dart
@@ -1,6 +1,7 @@
 library browser;
 
-export 'jaspr.dart' hide runApp;
+export 'jaspr.dart' hide runApp, AppContext;
+export 'src/browser/app_context.dart';
 export 'src/browser/browser_binding.dart';
 export 'src/browser/clients.dart';
 export 'src/browser/dom_render_object.dart';