+++ packages/jaspr/lib/src/server/run_app.dart
@@ -16,7 +16,10 @@ void runApp(Component app) {
   ServerApp.run(_createSetup(app));
 }
 
-/// Returns a shelf handler that serves the provided component and related assets
+typedef RenderFunction = FutureOr<Response> Function(Component);
+typedef AppHandler = FutureOr<Response> Function(Request, RenderFunction render);
+
+/// Returns a shelf handler that serves the provided component and related assets.
 Handler serveApp(AppHandler handler) {
   _checkInitialized('serveApp');
   return createHandler((request, render) {
@@ -26,19 +29,18 @@ Handler serveApp(AppHandler handler) {
   });
 }
 
-typedef RenderFunction = FutureOr<Response> Function(Component);
-typedef AppHandler = FutureOr<Response> Function(Request, RenderFunction render);
+/// A record containing the status code, body, and headers for a response.
+typedef ResponseLike = ({int statusCode, String body, Map<String, List<String>> headers});
 
-/// Directly renders the provided component into a html string.
+/// Directly renders the provided component to HTML. Returns a [ResponseLike] object.
 ///
-/// When [standalone] is false (default), the html output will have a full document structure (html, head, body).
-Future<String> renderComponent(Component app, {bool standalone = false}) async {
+/// - Accepts a [request] object for getting the current url and headers.
+/// - When [standalone] is false (default), the html output will have a full document structure (html, head, body).
+Future<ResponseLike> renderComponent(Component app, {Request? request, bool standalone = false}) async {
   _checkInitialized('renderComponent');
-  var fileHandler = staticFileHandler();
-  return render(_createSetup(app), Uri.parse('https://0.0.0.0/'), (name) async {
-    var response = await fileHandler(Request('get', Uri.parse('https://0.0.0.0/$name')));
-    return response.statusCode == 200 ? response.readAsString() : null;
-  }, standalone);
+  request ??= Request('GET', Uri.parse('https://0.0.0.0/'));
+  var fileLoader = proxyFileLoader(request, staticFileHandler());
+  return render(_createSetup(app), request, fileLoader, standalone);
 }
 
 void _checkInitialized(String method) {