+++ packages/jaspr_router/lib/src/platform/platform.dart
@@ -16,7 +16,7 @@ abstract class PlatformRouter {
 /// Will be implemented separately on browser and server
 abstract class HistoryManager {
   /// Initialize the history manager and setup any listeners to history changes
-  void init(AppBinding binding, {void Function(Object? state, {String? url})? onChangeState});
+  void init(BuildContext context, {void Function(Object? state, {String? url})? onChangeState});
 
   /// Push a new state to the history
   void push(String url, {String? title, Object? data});